requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==24.2.0
black==24.10.0
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.5
cymem==2.0.8
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.2
fastapi==0.104.1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
groq==0.12.0
grpcio==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langcodes==3.4.1
langsmith==0.1.144
language_data==1.2.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.24.3
openai==1.54.4
orjson==3.10.11
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
portalocker==2.10.1
preshed==3.0.9
propcache==0.2.0
protobuf==5.28.3
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv==1.0.0
PyYAML==6.0.2
qdrant-client==1.6.4
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.11.3
sentence-transformers==3.3.0
sentencepiece==0.2.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
starlette==0.27.0
sympy==1.13.1
tenacity==9.0.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tomli==2.1.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers==4.46.2
typer==0.9.4
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.24.0
wasabi==1.1.3
weasel==0.3.4
yarl==1.18.0
